requirements.txt
boto3==1.26.151
botocore==1.29.151
CacheControl==0.13.1
certifi==2022.12.7
charset-normalizer==3.1.0
dill==0.3.6
idna==3.4
iso3166==2.1.1
jmespath==1.0.1
joblib==1.2.0
mapbox==0.18.1
msgpack==1.0.5
numpy==1.24.3
polyline==2.0.0
progress==1.6
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
regex==2023.6.3
requests==2.28.2
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.15
